Fixing use of multiple patterns; now use one pattern for one log
It turns out I had completely mistaken the intent of the ability
to list multiple regular expressions.  I had thought it was a
convenient way of listing different ways of interpreting the logs
within one file.  This is actually for recognising different files
- either different programs or different versions of the same
program.

What this means is that lnav looks at the file and decides on the
most appropriate _regex_ to choose based on how many matches it
gets.  It uses no other regexes - from that format file or anywhere
else -  even if they would match.

So I've redone all my log formats to use only one format for
interpreting logs from the same program.  In most cases this makes
the regex much more complicated, but there's no helping that.