Considerable fixes to regex to get all the parts of an strace line
Because the actual syscall appears in different places in different situations,
we now split that into syscall, syscall_unfinished and syscall_resumed.  We
try really hard to match the arguments, including structures, and the return
code, error name, error description and duration sections.  All optionally.

Note that this format asserts the existence of a timestamp, because straces
without timestamps are not very interesting to lnav.